sbt and shader groups look ok

+++ Shaders/SimpleRaytracing/GeometryNode.glsl
@@ -1,7 +1,8 @@
+// TODO Probably delete this
 struct GeometryNode
 {
-	uint64_t vertexBufferDeviceAddress;
-	uint64_t indexBufferDeviceAddress;
+	uint vertexBufferDeviceAddress;
+	uint indexBufferDeviceAddress;
 	int textureIndexBaseColor;
 	int textureIndexOcclusion;
 };
